+++ modules/carto/src/index.ts
@@ -1,6 +1,6 @@
 export {getDefaultCredentials, setDefaultCredentials} from './config';
 export {default as CartoLayer} from './layers/carto-layer';
-export {default as _CartoDynamicTileLayer} from './layers/carto-dynamic-tile-layer';
+export {default as _CartoTileLayer} from './layers/carto-tile-layer';
 export {default as BASEMAP} from './basemap';
 export {default as colorBins} from './style/color-bins-style';
 export {default as colorCategories} from './style/color-categories-style';
